+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/signaling/SignalingConfig.java
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.android.communication.chat.signaling;
+
+class SignalingConfig {
+
+    public static final String TROUTER_CLIENT_VERSION = "99116/0.0.0.0/";
+    public static final String DEFAULT_TROUTER_HOSTNAME = "go.trouter-int.skype.net";
+    public static final String DEFAULT_REGISTRATION_HOSTNAME_AND_BASE_PATH = "edge.skype.net/registrar/testenv";
+    public static final int DEFAULT_MAX_REGISTRATION_TTLS = 3600;
+
+    public static final String TEST_APPLICATION_ID = "AcsAndroid_test";
+    public static final String TEST_TEMPLATE_KEY = "AcsAndroid_Chat_test_1.1";
+    public static final String PLATFORM = "SPOOL";
+    public static final String PLATFORM_UI_VERSION = "0.0.0.0";
+
+
+}