+++ sdk/communication/azure-communication-common/src/test/java/com/azure/android/communication/common/CommunicationTokenCredentialTest.java
@@ -39,8 +39,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
 
         expectedException.expect(is(instanceOf(IllegalArgumentException.class)));
-
-        new CommunicationTokenCredential(mockTokenRefresher, "This is an invalid token string");
+        new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, "This is an invalid token string"));
     }
 
     @Test
@@ -51,7 +50,7 @@ public class CommunicationTokenCredentialTest {
         CountDownLatch countDownLatch = new CountDownLatch(1);
         mockTokenRefresher.setOnCallReturn(countDownLatch::countDown);
 
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, true);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
         countDownLatch.await();
 
         CommunicationAccessToken accessToken = credential.getToken().get();
@@ -67,7 +66,7 @@ public class CommunicationTokenCredentialTest {
         CountDownLatch countDownLatch = new CountDownLatch(1);
         mockTokenRefresher.setOnCallReturn(countDownLatch::countDown);
 
-        new CommunicationTokenCredential(mockTokenRefresher, true, tokenString);
+        new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true, tokenString));
         countDownLatch.await();
 
         assertEquals(1, mockTokenRefresher.getCallCount());
@@ -78,7 +77,7 @@ public class CommunicationTokenCredentialTest {
         String tokenString = TokenStubHelper.createTokenStringForOffset(700);
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
 
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, true, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true, tokenString));
         CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertEquals(tokenString, accessToken.getToken());
@@ -89,7 +88,7 @@ public class CommunicationTokenCredentialTest {
     public void constructor_withTokenRefresher_proactiveRefresh_withInitialTokenInvalid() {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         expectedException.expect(is(instanceOf(IllegalArgumentException.class)));
-        new CommunicationTokenCredential(mockTokenRefresher, true, "This is an invalid token string");
+        new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true, "This is an invalid token string"));
     }
 
     @Test
@@ -111,7 +110,7 @@ public class CommunicationTokenCredentialTest {
             countDownLatch.countDown();
         });
 
-        new CommunicationTokenCredential(mockTokenRefresher, true);
+        new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
         countDownLatch.await();
 
         assertEquals(numRepeats, mockTokenRefresher.getCallCount());
@@ -162,7 +161,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(300);
         mockTokenRefresher.setToken(refreshedToken);
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false));
 
         CommunicationAccessToken accessToken = credential.getToken().get();
 
@@ -180,7 +179,7 @@ public class CommunicationTokenCredentialTest {
         mockTokenRefresher.setOnCallReturn(() -> {
             throw mockTokenRefresherException;
         });
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false));
 
         try {
             expectedException.expectCause(is(mockTokenRefresherException));
@@ -195,7 +194,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(300);
         mockTokenRefresher.setToken(refreshedToken);
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false));
 
         // Set up blocked multithreaded calls
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
@@ -223,7 +222,7 @@ public class CommunicationTokenCredentialTest {
     public void getToken_onDemandAutoRefresh_tokenWithinThresholdNotRefreshed_singleCall() throws ExecutionException, InterruptedException {
         String tokenString = TokenStubHelper.createTokenStringForOffset(130);
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
         CommunicationAccessToken accessToken = credential.getToken().get();
 
@@ -235,7 +234,7 @@ public class CommunicationTokenCredentialTest {
     public void getToken_onDemandAutoRefresh_tokenWithinThresholdNotRefreshed_multithreadedCalls() throws ExecutionException, InterruptedException {
         String tokenString = TokenStubHelper.createTokenStringForOffset(130);
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
         // Set up blocked multithreaded calls
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
@@ -266,7 +265,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(300);
         mockTokenRefresher.setToken(refreshedToken);
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
         CommunicationAccessToken accessToken = credential.getToken().get();
 
         assertNotEquals(tokenString, accessToken.getToken());
@@ -282,7 +281,7 @@ public class CommunicationTokenCredentialTest {
         mockTokenRefresher.setOnCallReturn(() -> {
             throw mockTokenRefresherException;
         });
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
         try {
             expectedException.expectCause(is(mockTokenRefresherException));
@@ -298,7 +297,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(300);
         mockTokenRefresher.setToken(refreshedToken);
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
         Set<CommunicationAccessToken> accessTokens = new HashSet<>();
         int numCalls = 3;
@@ -319,7 +318,7 @@ public class CommunicationTokenCredentialTest {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(300);
         mockTokenRefresher.setToken(refreshedToken);
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
 
         // Set up blocked multithreaded calls
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
@@ -348,7 +347,7 @@ public class CommunicationTokenCredentialTest {
     @Test
     public void getToken_onDemandAutoRefresh_isDisposed_cancelledFuture() throws ExecutionException, InterruptedException {
         MockTokenRefresher mockTokenRefresher = new MockTokenRefresher();
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false));
 
         credential.dispose();
         Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
@@ -365,7 +364,7 @@ public class CommunicationTokenCredentialTest {
         String refreshedToken = TokenStubHelper.createTokenStringForOffset(1200);
         mockTokenRefresher.setToken(refreshedToken);
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, true);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
 
         Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
         blockedRefresh.run();
@@ -384,7 +383,7 @@ public class CommunicationTokenCredentialTest {
         mockTokenRefresher.setToken(refreshedToken);
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
 
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, tokenString);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, false, tokenString));
         Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
         credential.dispose();
         blockedRefresh.run();
@@ -403,7 +402,7 @@ public class CommunicationTokenCredentialTest {
         mockTokenRefresher.setToken(refreshedToken);
         Runnable blockedRefresh = this.arrangeBlockedRefresh(mockTokenRefresher);
 
-        CommunicationTokenCredential credential = new CommunicationTokenCredential(mockTokenRefresher, true);
+        CommunicationTokenCredential credential = new CommunicationTokenCredential(new CommunicationTokenRefreshOptions(mockTokenRefresher, true));
         Future<CommunicationAccessToken> accessTokenFuture = credential.getToken();
         credential.dispose();
         blockedRefresh.run();